* emacs-22.3-appkit-1.11 (2010-09-23)
** Fixed bugs

*** Boundary indicators in fringes are scrolled out if vscrolled.
Apply a fix for Bug#5634 and Bug#6325.

*** Prefix keys defined in ~/Library/KeyBindings/DefaultKeyBinding.dict
need to be typed multiple times.

*** Successive paste of different images from other applications via
the clipboard only inserts the first image.

*** 64-bit binary built on Mac OS X 10.5 does not run on 10.6.

*** Can't quit while establishing a TCP connection.
Apply a fix for Bug#5723 as well as Helmut Eller's fix for Bug#5173.

** Improvements

*** "Speak selected text with when the key is pressed", which can be
customized in the Speech pane in the System Preferences, now works.

*** Very experimental support for mouse wheel smooth scroll.
Note that it still has several glitches and you can turn it off by
setting `mac-mouse-wheel-smooth-scroll' to nil.

*** Menu item "Open Selected File in Emacs" is shown in Services or
context menu of other applications by default on Mac OS X 10.6 when
absolute pathname-like text is selected.

*** "Click in the scroll bar to: Jump to the spot that's clicked"
setting in the System Preferences is supported.